backend/requirements.txt
torch
torchvision
fastapi
uvicorn
pydantic
numpy
pandas
pillow
scikit-learn
opencv-python
matplotlib
tqdm
requests
python-multipart  # Needed if handling file uploads in FastAPI
